Refactored logger
