+++ docs/documentation/logger.rst
@@ -1,0 +1,15 @@
+Logger
+=========================================
+
+The ``quantum-robot`` package stores logging information from
+qUnits subprocesses and log them in text files.
+
+.. note:: 
+    The logs are stored in a ``.qrobot_logs`` folder
+    in the current working directory in which the main python
+    process is executed
+
+.. automodule:: qrobot._logger
+   :members: log_dir, log_file
+
+
