--- src/qrobot/logs/__init__.py
@@ -1,3 +1,0 @@
-from .logger import get_logger
-
-__all__ = ["get_logger"]
